Refactor model field definitions to allow null values and adjust integer field types
- Updated various model fields across multiple files to allow null values where appropriate, enhancing data integrity and flexibility.
- Changed specific integer fields to BigIntegerField to accommodate larger values.
- Removed unnecessary migration file and consolidated changes into existing migrations.
- Adjusted the extract_fields method to determine the appropriate integer field type based on value ranges.
- Cleaned up the convert-sde-yaml-to-json.sh script by removing redundant file deletion commands.

+++ apps/sde/models/stars.py
@@ -6,6 +6,6 @@ from django.db import models
 
 class Star(models.Model):
     id = models.IntegerField(primary_key=True)
-    radius = models.IntegerField()
+    radius = models.BigIntegerField()
     statistics = models.JSONField()
     type_id = models.IntegerField()
